my_copilot/requirements.txt
streamlit==1.32.0
groq
huggingface_hub
python-dotenv
requests
pillow
